Extract `dispute` tests
