+++ zrml/prediction-markets/src/old_tests.rs
@@ -138,177 +138,6 @@ fn simple_create_scalar_market(
 }
 
 #[test]
-fn it_allows_to_dispute_the_outcome_of_a_market() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 2;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-
-        // Run to the end of the trading phase.
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-
-        let dispute_at = grace_period + 2;
-        run_to_block(dispute_at);
-
-        assert_ok!(PredictionMarkets::dispute(RuntimeOrigin::signed(CHARLIE), 0,));
-        assert_ok!(SimpleDisputes::suggest_outcome(
-            RuntimeOrigin::signed(CHARLIE),
-            0,
-            OutcomeReport::Categorical(0)
-        ));
-
-        let market = MarketCommons::market(&0).unwrap();
-        assert_eq!(market.status, MarketStatus::Disputed);
-
-        let disputes = zrml_simple_disputes::Disputes::<Runtime>::get(0);
-        assert_eq!(disputes.len(), 1);
-        let dispute = &disputes[0];
-        assert_eq!(dispute.at, dispute_at);
-        assert_eq!(dispute.by, CHARLIE);
-        assert_eq!(dispute.outcome, OutcomeReport::Categorical(0));
-
-        let dispute_ends_at = dispute_at + market.deadlines.dispute_duration;
-        let market_ids = MarketIdsPerDisputeBlock::<Runtime>::get(dispute_ends_at);
-        assert_eq!(market_ids.len(), 1);
-        assert_eq!(market_ids[0], 0);
-    });
-}
-
-#[test]
-fn dispute_fails_disputed_already() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 2;
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Block(0..end),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(<Runtime as Config>::MinCategories::get()),
-            Some(MarketDisputeMechanism::Authorized),
-            ScoringRule::Lmsr,
-        ));
-
-        // Run to the end of the trading phase.
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-
-        let dispute_at = grace_period + 2;
-        run_to_block(dispute_at);
-
-        assert_ok!(PredictionMarkets::dispute(RuntimeOrigin::signed(CHARLIE), 0,));
-
-        assert_noop!(
-            PredictionMarkets::dispute(RuntimeOrigin::signed(CHARLIE), 0),
-            Error::<Runtime>::InvalidMarketStatus,
-        );
-    });
-}
-
-#[test]
-fn dispute_fails_if_market_not_reported() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 2;
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Block(0..end),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(<Runtime as Config>::MinCategories::get()),
-            Some(MarketDisputeMechanism::Authorized),
-            ScoringRule::Lmsr,
-        ));
-
-        // Run to the end of the trading phase.
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-
-        // no report happening here...
-
-        let dispute_at = grace_period + 2;
-        run_to_block(dispute_at);
-
-        assert_noop!(
-            PredictionMarkets::dispute(RuntimeOrigin::signed(CHARLIE), 0),
-            Error::<Runtime>::InvalidMarketStatus,
-        );
-    });
-}
-
-#[test]
-fn dispute_reserves_dispute_bond() {
-    ExtBuilder::default().build().execute_with(|| {
-        let end = 2;
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Block(0..end),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(<Runtime as Config>::MinCategories::get()),
-            Some(MarketDisputeMechanism::Authorized),
-            ScoringRule::Lmsr,
-        ));
-
-        // Run to the end of the trading phase.
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            0,
-            OutcomeReport::Categorical(1)
-        ));
-
-        let dispute_at = grace_period + 2;
-        run_to_block(dispute_at);
-
-        let free_charlie_before = Balances::free_balance(CHARLIE);
-        let reserved_charlie = Balances::reserved_balance(CHARLIE);
-        assert_eq!(reserved_charlie, 0);
-
-        assert_ok!(PredictionMarkets::dispute(RuntimeOrigin::signed(CHARLIE), 0,));
-
-        let free_charlie_after = Balances::free_balance(CHARLIE);
-        assert_eq!(free_charlie_before - free_charlie_after, DisputeBond::get());
-
-        let reserved_charlie = Balances::reserved_balance(CHARLIE);
-        assert_eq!(reserved_charlie, DisputeBond::get());
-    });
-}
-
-#[test]
 fn schedule_early_close_emits_event() {
     ExtBuilder::default().build().execute_with(|| {
         let end = 100;
